Move common string utils to a module
